ar71xx: Fix clock workaround for TP-Link CPE210v2
The CPE210v2 workaround must not change the behavior of other devices. The
exception must therefore only be applied when the TP-Link CPE210v2 was
detected.

The refclk is done when mips_get_machine_name still returns "Unknown". It
is therefore required to use mips_machtype instead.

Also, the return code of strcmp for "equal" is 0 and not != 0.

Signed-off-by: Sven Eckelmann <sven@narfation.org>